Move abstract game engine to a new package
