+++ src/main/java/fr/nc0/cda/controleur/ControleurTemplate.java
@@ -6,9 +6,9 @@
 
 package fr.nc0.cda.controleur;
 
-import fr.nc0.cda.modele.CoupInvalideException;
-import fr.nc0.cda.modele.EtatPartie;
-import fr.nc0.cda.modele.EtatPartieException;
+import fr.nc0.cda.modele.jeu.CoupInvalideException;
+import fr.nc0.cda.modele.jeu.EtatPartie;
+import fr.nc0.cda.modele.jeu.EtatPartieException;
 import fr.nc0.cda.modele.joueur.Joueur;
 import fr.nc0.cda.vue.Ihm;
