+++ src/main/java/fr/nc0/cda/modele/jeu/CoupInvalideException.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.jeu;
 
 /** Exception levée lorsqu'un coup est invalide. */
 public class CoupInvalideException extends Exception {
