+++ src/main/java/fr/nc0/cda/modele/jeu/EtatPartie.java
@@ -4,7 +4,7 @@
  * that can be found in the COPYRIGHT file.
  */
 
-package fr.nc0.cda.modele;
+package fr.nc0.cda.modele.jeu;
 
 /** Représente les états possibles d'une partie à deux joueurs. */
 public enum EtatPartie {
