+++ src/main/java/fr/nc0/cda/modele/nim/PlateauNim.java
@@ -6,7 +6,7 @@
 
 package fr.nc0.cda.modele.nim;
 
-import fr.nc0.cda.modele.Plateau;
+import fr.nc0.cda.modele.jeu.Plateau;
 import java.util.ArrayList;
 import java.util.List;
